backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic-settings==2.5.0
PyJWT==2.9.0
redis[asyncio]==5.0.8
groq==0.11.0
python-multipart==0.0.12
boto3==1.35.0
httpx==0.27.0
pytesseract==0.3.13
Pillow==10.4.0
python-dateutil==2.9.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
requests==2.32.0
fpdf2==2.8.7
